Added some example tweet searching - after successful login, searches for cats.
Temporary: clearing active Twitter session, so that re-login exercises the query again.

Toasts are used to display query start and query result.  Note that an AsyncTask is used because network activity is (quite reasonably) not allowed on UI thread.

Added LocationTracker - if we want live location updates, that'll be where to find 'em.